requirements.txt
# Deep Learning Framework
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# LLM and RAG
langchain>=0.0.200
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
chromadb>=0.4.0

# RL and Optimization
gymnasium>=0.28.1
stable-baselines3>=2.1.0
trl>=0.7.1  # For RLHF
peft>=0.4.0  # For parameter efficient fine-tuning

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Web and API
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0

# Testing and Development
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
loguru>=0.7.0 